Add commit command and make changes to package description
The commit cmd is added to the project, which includes all the functionality previous included in the `main` method.  Creates the initial groundwork to support additional commands
outside of generating commit messages with coco.

In addition, the package description in the package.json file is modified to "zero-effort git commits with coco."
